requirements.txt
gradio
pandas
plotly
u8darts
openpyxl
